+++ src/implementation/post.rs
@@ -3,8 +3,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 use crate::implementation::{ContractMode, ContractType, FuncWithContracts};
-use proc_macro::TokenStream;
-use syn::ItemFn;
+use proc_macro2::TokenStream;
 
 pub(crate) fn post(
     mode: ContractMode,
@@ -13,7 +12,7 @@ pub(crate) fn post(
 ) -> TokenStream {
     let ty = ContractType::Post;
 
-    let func = syn::parse_macro_input!(toks as ItemFn);
+    let func = syn::parse_quote!(#toks);
 
     let f = FuncWithContracts::new_with_initial_contract(func, ty, mode, attr);